// src/modules/service/service.interface.d.ts
export interface ICreateService {
    title: string;
    description: string;
    category: string;
    price: number;
    image?: string;
}
export interface IUpdateService {
    title?: string;
    description?: string;
    category?: string;
    price?: number;
    image?: string;
    isActive?: boolean;
}
//# sourceMappingURL=service.interface.d.ts.map